requirements.txt
asgiref==3.9.1
defusedxml==0.7.1
Django==5.2.4
django-widget-tweaks==1.5.0
fonttools==4.59.0
fpdf==1.7.2
numpy==2.3.1
pandas==2.3.1
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
gunicorn==23.0.0